refac: rename PatchExtractor to DataReader and remove patch generation logic
